requirements.txt
amqp==2.4.1
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
kombu==4.3.0
MarkupSafe==1.1.0
numpy==1.16.1
Pillow==5.4.1
pytz==2018.9
redis==3.2.0
requests==2.21.0
six==1.12.0
torch==1.0.1.post2
torchvision==0.2.1
urllib3==1.24.1
vine==1.2.0
Werkzeug==0.14.1
